MVP Architecture in GroupsListFragment
