+++ mifosng-android/src/main/java/com/mifos/api/DataManager.java
@@ -99,7 +99,7 @@ public class DataManager {
 
     public Observable<SaveResponse> saveCollectionSheetAsync(int id,
                                                              CollectionSheetPayload payload) {
-        return mBaseApiManager.getCenterApi().saveCollectionSheetAsync(id,payload);
+        return mBaseApiManager.getCenterApi().saveCollectionSheetAsync(id, payload);
     }
 
     /**
@@ -185,6 +185,13 @@ public class DataManager {
         return mBaseApiManager.getGroupApi().getGroup(groupId);
     }
 
+    public Observable<Page<Group>> getAllGroup() {
+        return mBaseApiManager.getGroupApi().getAllGroup();
+    }
+
+    public Observable<Page<Group>> listAllGroups(int offset, int limit) {
+        return mBaseApiManager.getGroupApi().listAllGroups(offset,limit);
+    }
 
     /**
      * Offices API
