+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/injection/component/ActivityComponent.java
@@ -19,6 +19,7 @@ import com.mifos.mifosxdroid.online.documentlist.DocumentListFragment;
 import com.mifos.mifosxdroid.online.generatecollectionsheet.GenerateCollectionSheetFragment;
 import com.mifos.mifosxdroid.online.groupdetails.GroupDetailsFragment;
 import com.mifos.mifosxdroid.online.grouplist.GroupListFragment;
+import com.mifos.mifosxdroid.online.groupslist.GroupsListFragment;
 import com.mifos.mifosxdroid.online.loanaccountsummary.LoanAccountSummaryFragment;
 import com.mifos.mifosxdroid.online.loancharge.LoanChargeFragment;
 import com.mifos.mifosxdroid.online.savingaccountsummary.SavingsAccountSummaryFragment;
@@ -78,4 +79,6 @@ public interface ActivityComponent {
     void inject(CollectionSheetFragment collectionSheetFragment);
 
     void inject(SaveOfflineDataHelper saveOfflineDataHelper);
+
+    void inject(GroupsListFragment groupsListFragment);
 }
